Moved styled components to lib
