+++ frontend/src/lib/Container.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import styled from 'styled-components'
+
+export const Container = styled.div`
+height: 100vh;
+display: flex;
+flex-direction: row;
+justify-content: center;
+align-items: center;
+background: #F9E9FA;
+`;
