+++ frontend/src/lib/CreateUserStyle.js
@@ -1,0 +1,26 @@
+import React from 'react'
+import styled from 'styled-components'
+
+export const CreateUserContainer = styled.div`
+  border-radius: 0 20px 20px 0;
+  width: 35%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  background: white;
+  box-shadow: 5px 5px 5px grey;
+`;
+
+export const Register = styled.form`
+  display: flex;
+  flex-direction: column;
+  align-items:center;
+  width: 300px;
+`;
+
+export const Title = styled.h1`
+  align-self: flex-start;
+  color: purple;
+  font-size: 36px;
+  padding: 20px 0;
+`;
